refactor:TOP-65 feature:tohot 모듈 패키지 정리
