+++ feature/tohot/src/main/java/tht/feature/tohot/navigation/ToHotNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import tht.feature.tohot.screen.ToHotRoute
+import tht.feature.tohot.tohot.route.ToHotRoute
 
 @Composable
 fun ToHotNavigation() {
