+++ feature/tohot/src/main/java/tht/feature/tohot/tohot/screen/TopicWaitScreen.kt
@@ -1,4 +1,4 @@
-package tht.feature.tohot.screen
+package tht.feature.tohot.tohot.screen
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
